Implement field creation functionality: add createField action to handle field data submission, integrate CreateField component in facilities page, and remove obsolete createField route files.
